phase 2: rewrite experiment imports to reach ../shared/ after move
